backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
redis
pymysql